Make shape as a signed vector in SimpleArray

+++ cpp/solvcon/linalg/factorization.hpp
@@ -82,9 +82,9 @@ Llt<T>::array_type Llt<T>::forward_substitution(array_type const & l, array_type
         throw std::invalid_argument(oss.str());
     }
 
-    const auto m = static_cast<ssize_t>(l.shape(0));
-    const auto n = static_cast<ssize_t>(b.shape(1));
-    small_vector<size_t> const y_shape{static_cast<size_t>(m), static_cast<size_t>(n)};
+    const auto m = l.shape(0);
+    const auto n = b.shape(1);
+    small_vector<ssize_t> const y_shape{m, n};
     array_type y(y_shape);
     for (ssize_t k = 0; k < n; ++k)
     {
@@ -135,9 +135,9 @@ Llt<T>::array_type Llt<T>::backward_substitution(array_type const & l, array_typ
         throw std::invalid_argument(oss.str());
     }
 
-    const auto m = static_cast<ssize_t>(l.shape(0));
-    const auto n = static_cast<ssize_t>(y.shape(1));
-    small_vector<size_t> const x_shape{static_cast<size_t>(m), static_cast<size_t>(n)};
+    const auto m = l.shape(0);
+    const auto n = y.shape(1);
+    small_vector<ssize_t> const x_shape{m, n};
     array_type x(x_shape);
     for (ssize_t k = 0; k < n; ++k)
     {
@@ -173,8 +173,8 @@ Llt<T>::array_type Llt<T>::factorize(array_type const & a)
         throw std::invalid_argument(oss.str());
     }
 
-    const auto m = static_cast<ssize_t>(a.shape(0));
-    small_vector<size_t> const shape = {static_cast<size_t>(m), static_cast<size_t>(m)};
+    const auto m = a.shape(0);
+    small_vector<ssize_t> const shape = {m, m};
     array_type l(shape, value_type(0));
     const real_type eps = std::numeric_limits<real_type>::epsilon();
     for (ssize_t i = 0; i < m; ++i)
@@ -241,10 +241,10 @@ Llt<T>::array_type Llt<T>::solve(array_type const & a, array_type const & b)
     bool const was_1d = (b.ndim() == 1);
     if (was_1d)
     {
-        array_type const b_2d = b.reshape(small_vector<size_t>{b.shape(0), 1});
+        array_type const b_2d = b.reshape(small_vector<ssize_t>{b.shape(0), 1});
         array_type const y = forward_substitution(l, b_2d);
         array_type const x = backward_substitution(l, y);
-        return x.reshape(small_vector<size_t>{x.shape(0)});
+        return x.reshape(small_vector<ssize_t>{x.shape(0)});
     }
     else
     {
